// src/api/v1/services/enums/service.enum.ts
export enum EServicePaymentTypes {
	Percent = 'percent',
	Amount = 'amount',
}

export enum EServiceTypes {
	Rent = 'rent',
	Additional = 'additional',
}

export enum EServiceFormats {
	Time = 'time',
	TimeAndSeat = 'timeAndSeat',
	Seat = 'seat',
	Free = 'free',
}

export enum EServicePrepayment {
	Without = 'without',
	Main = 'main',
	MainAndAdditional = 'mainAndAdditional',
}

export enum EBlockBookingType {
	MIN = 'min',
	HOUR = 'hour',
}

export enum EPaymentTypes {
	TIME = 'time',
	PIECE = 'piece',
}

export enum EAdditionalTypes {
	REQUISITE = 'requisite',
	PROCESS = 'process',
}